test: cleanup and refactor
