+++ FountainPensNg.xTests/FountainPenTests.cs
@@ -1,16 +1,19 @@
 namespace FountainPensNg.xTests;
 
+[CollectionDefinition("FountainPen Tests")]
+public class FountainPenCollection : ICollectionFixture<InkedUpModuleFixture>;
+
 [Collection("FountainPen Tests")]
 public class FountainPenTests {
 
-	private readonly FountainPenModuleFixture _fixture;
-	public FountainPenTests(FountainPenModuleFixture fixture) {
+	private readonly InkedUpModuleFixture _fixture;
+	public FountainPenTests(InkedUpModuleFixture fixture) {
 		_fixture = fixture;
 	}
 
 	[Fact]
 	public async Task GetFountainPen() {
-		var pen = _fixture.TestSeed.FountainPens[0];
+		var pen = _fixture.FountainPens[0];
 		var client = _fixture.Factory.CreateClient();
 		var response = await client.GetAsync($"/api/fountain-pens/{pen.Id}");
 
@@ -45,7 +48,7 @@ public class FountainPenTests {
 	[Fact]
 	public async Task UpdateFountainPen() {
 		var client = _fixture.Factory.CreateClient();
-		var pen = _fixture.TestSeed.FountainPens[0];
+		var pen = _fixture.FountainPens[0];
 		var dto = pen.Adapt<FountainPenUploadDTO>();
 		var content = JsonContent.Create(dto);
 		var response = await client.PutAsync($"/api/fountain-pens/{dto.Id}", content);
@@ -57,7 +60,7 @@ public class FountainPenTests {
 
 	[Fact]
 	public async Task DeleteFountainPen() {
-		var pen = _fixture.TestSeed.FountainPens[1];
+		var pen = _fixture.FountainPens[1];
 		var client = _fixture.Factory.CreateClient();
 		var response = await client.DeleteAsync($"/api/fountain-pens/{pen.Id}");
 		response.EnsureSuccessStatusCode();
@@ -81,7 +84,7 @@ public class FountainPenTests {
 	[Fact]
 	public async Task GetFullText() {
 		var client = _fixture.Factory.CreateClient();
-		var response = await client.GetAsync($"/api/finder/{_fixture.TestSeed.FountainPens[0].ModelName}");
+		var response = await client.GetAsync($"/api/finder/{_fixture.FountainPens[0].ModelName}");
 		response.EnsureSuccessStatusCode();
 
 		var fountainPens = await response.Content.ReadFromJsonAsync<IEnumerable<SearchResultDTO>>();
